+++ DTO/CauHoiDTO.cs
@@ -11,14 +11,14 @@ namespace DTO
         public int MaCauHoi { get; set; }
         public string NoiDung { get; set; }
         public int MaMonHoc { get; set; }
-        public int MaNguoiTao { get; set; }
+        public long MaNguoiTao { get; set; }
         public string DoKho { get; set; }
         public int TrangThai { get; set; }
         public int is_delete { get; set; }
         public CauHoiDTO()
         {
         }
-        public CauHoiDTO(int maCauHoi, string noiDung, int maMonHoc, int maNguoiTao, string doKho, int trangThai, int is_delete)
+        public CauHoiDTO(int maCauHoi, string noiDung, int maMonHoc, long maNguoiTao, string doKho, int trangThai, int is_delete)
         {
             MaCauHoi = maCauHoi;
             NoiDung = noiDung;
